# zeus_openapi.json
{"components":{"schemas":{"ComponentStock":{"properties":{"category":{"type":"string"},"name":{"type":"string"},"reorder_point":{"type":"integer"},"sku":{"type":"string"},"status":{"enum":["In Stock","Low Stock","Out of Stock","Discontinued"],"type":"string"},"stock_qty":{"type":"integer"},"unit_cost":{"format":"float","type":"number"}},"type":"object"},"Error":{"properties":{"error":{"type":"string"}},"type":"object"},"GRLineItem":{"properties":{"aging_label":{"type":"string"},"aging_sensitive":{"type":"boolean"},"defective_qty":{"nullable":true,"type":"integer"},"gr_id":{"type":"string"},"id":{"format":"uuid","type":"string"},"name":{"type":"string"},"ordered_qty":{"type":"integer"},"production_date":{"format":"date","nullable":true,"type":"string"},"received_qty":{"nullable":true,"type":"integer"},"sku":{"type":"string"}},"type":"object"},"GoodsReceipt":{"properties":{"arrival_date":{"format":"date-time","type":"string"},"id":{"example":"GR-2025-1","type":"string"},"lock_expires_at":{"format":"date-time","nullable":true,"type":"string"},"locked_by":{"nullable":true,"type":"string"},"operator_id":{"type":"string"},"po_ref":{"type":"string"},"status":{"enum":["Pending","Inspected","Complete","Discrepancy"],"type":"string"},"vendor_id":{"format":"uuid","type":"string"}},"type":"object"},"POLineItem":{"properties":{"description":{"type":"string"},"id":{"format":"uuid","type":"string"},"ordered_qty":{"type":"integer"},"po_id":{"type":"string"},"received_qty":{"type":"integer"},"sku":{"type":"string"},"unit_price":{"format":"float","type":"number"}},"type":"object"},"PaginationMeta":{"properties":{"limit":{"type":"integer"},"page":{"type":"integer"},"total_pages":{"type":"integer"},"total_rows":{"type":"integer"}},"type":"object"},"Part":{"properties":{"created_at":{"format":"date-time","type":"string"},"id":{"format":"uuid","type":"string"},"installation_date":{"format":"date-time","nullable":true,"type":"string"},"manufactured_date":{"format":"date-time","type":"string"},"part_catalog_id":{"format":"uuid","type":"string"},"part_condition_id":{"type":"integer"},"product_id":{"format":"uuid","nullable":true,"type":"string"},"removal_date":{"format":"date-time","nullable":true,"type":"string"},"scrapped_date":{"format":"date-time","nullable":true,"type":"string"},"serial_number":{"type":"string"},"updated_at":{"format":"date-time","type":"string"}},"type":"object"},"Product":{"properties":{"created_at":{"format":"date-time","type":"string"},"customer_id":{"format":"uuid","type":"string"},"id":{"format":"uuid","type":"string"},"product_model_code":{"type":"string"},"product_name":{"type":"string"},"serial_number":{"type":"string"},"updated_at":{"format":"date-time","type":"string"}},"type":"object"},"PurchaseOrder":{"properties":{"expected_delivery":{"format":"date-time","type":"string"},"id":{"example":"PO-2025-1","type":"string"},"line_items":{"items":{"$ref":"#/components/schemas/POLineItem"},"type":"array"},"payment_terms":{"type":"string"},"status":{"enum":["Draft","Approved","In Transit","Received","Partial","Void"],"type":"string"},"target_build":{"type":"string"},"total_value":{"format":"float","type":"number"},"vendor_id":{"format":"uuid","type":"string"}},"type":"object"},"Shipment":{"properties":{"carrier":{"type":"string"},"eta":{"format":"date-time","type":"string"},"id":{"example":"SHP-2025-1","type":"string"},"items":{"items":{"$ref":"#/components/schemas/ShipmentItem"},"type":"array"},"po_ref":{"type":"string"},"ship_date":{"format":"date-time","type":"string"},"status":{"enum":["Scheduled","In Transit","Delivered","Delayed"],"type":"string"},"supplier_id":{"format":"uuid","type":"string"},"tracking_no":{"type":"string"}},"type":"object"},"ShipmentItem":{"properties":{"description":{"type":"string"},"id":{"format":"uuid","type":"string"},"qty":{"type":"integer"},"shipment_id":{"type":"string"},"sku":{"type":"string"}},"type":"object"},"SkuMapping":{"properties":{"id":{"format":"uuid","type":"string"},"lead_time_days":{"type":"integer"},"min_order_qty":{"type":"integer"},"name":{"type":"string"},"sku":{"type":"string"},"supplier_id":{"format":"uuid","type":"string"},"unit_price":{"format":"float","type":"number"}},"type":"object"},"SuccessListResponse":{"properties":{"data":{"properties":{"items":{"type":"array"},"pagination":{"$ref":"#/components/schemas/PaginationMeta"}},"type":"object"},"message":{"example":"success","type":"string"},"statusCode":{"example":200,"type":"integer"}},"type":"object"},"SuccessResponse":{"properties":{"data":{"nullable":true},"message":{"example":"success","type":"string"},"statusCode":{"example":200,"type":"integer"}},"type":"object"},"Supplier":{"properties":{"contact":{"type":"string"},"id":{"format":"uuid","type":"string"},"lead_time_days":{"type":"integer"},"name":{"type":"string"},"on_time_rate":{"format":"float","type":"number"},"quality_score":{"format":"float","type":"number"},"tier":{"enum":["Preferred","Qualified","Under Review"],"type":"string"}},"type":"object"}},"securitySchemes":{"BearerAuth":{"bearerFormat":"JWT","scheme":"bearer","type":"http"}}},"info":{"description":"Supply Chain Management module of the Zeus ERP system. Governs procurement, inventory, goods receipt, and outbound logistics.","title":"Zeus SCM API","version":"1.0.0"},"openapi":"3.0.3","paths":{"/goods-receipts/{grId}/lock":{"delete":{"parameters":[{"in":"path","name":"grId","required":true,"schema":{"type":"string"}}],"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SuccessResponse"}}},"description":"Lock released"}},"summary":"Release 60-min operator lock on Goods Receipt","tags":["Goods Receipts"]},"post":{"parameters":[{"in":"path","name":"grId","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"properties":{"operator_id":{"type":"string"}},"required":["operator_id"],"type":"object"}}},"required":true},"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SuccessResponse"}}},"description":"Lock acquired"},"409":{"description":"Already locked by another operator"}},"summary":"Acquire 60-min operator lock on Goods Receipt","tags":["Goods Receipts"]}},"/goods-receipts/{grId}/process":{"post":{"parameters":[{"in":"path","name":"grId","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"properties":{"counts":{"additionalProperties":{"properties":{"defective":{"minimum":0,"type":"integer"},"received":{"minimum":0,"type":"integer"}},"type":"object"},"type":"object"},"operator_id":{"type":"string"}},"required":["operator_id","counts"],"type":"object"}}},"required":true},"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SuccessResponse"}}},"description":"Blind receipt processed"}},"summary":"Process blind receipt with manual counts","tags":["Goods Receipts"]}},"/health":{"get":{"responses":{"200":{"description":"OK"}},"security":[],"summary":"Health check","tags":["System"]}},"/inventory/part-catalog":{"get":{"parameters":[{"description":"Filter by part type ID","in":"query","name":"type_id","schema":{"type":"integer"}},{"description":"Page number","in":"query","name":"page","schema":{"default":1,"type":"integer"}},{"description":"Items per page (max 100)","in":"query","name":"limit","schema":{"default":15,"type":"integer"}},{"description":"Sort column","in":"query","name":"sort_by","schema":{"default":"created_at","enum":["created_at","updated_at","part_number","mfg_number"],"type":"string"}},{"description":"Sort direction","in":"query","name":"sort_dir","schema":{"default":"desc","enum":["asc","desc"],"type":"string"}},{"description":"Search keyword (matches part_number, mfg_number, description)","in":"query","name":"q","schema":{"type":"string"}}],"responses":{"200":{"content":{"application/json":{"schema":{"allOf":[{"$ref":"#/components/schemas/SuccessResponse"},{"properties":{"data":{"properties":{"items":{"items":{"$ref":"#/components/schemas/Part"},"type":"array"},"pagination":{"$ref":"#/components/schemas/PaginationMeta"}},"type":"object"}}}]}}},"description":"Part catalog list"}},"summary":"List part catalog with filtering, pagination, sorting, and search","tags":["Inventory"]}},"/inventory/part-catalog/{id}":{"get":{"parameters":[{"in":"path","name":"id","required":true,"schema":{"format":"uuid","type":"string"}}],"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SuccessResponse"}}},"description":"Part catalog entry"}},"summary":"Get part catalog entry by ID","tags":["Inventory"]}},"/inventory/parts":{"get":{"parameters":[{"description":"Filter by part catalog ID","in":"query","name":"catalog_id","schema":{"format":"uuid","type":"string"}},{"description":"Filter by product ID","in":"query","name":"product_id","schema":{"format":"uuid","type":"string"}},{"description":"Filter by part condition ID","in":"query","name":"condition_id","schema":{"type":"integer"}},{"description":"Page number","in":"query","name":"page","schema":{"default":1,"type":"integer"}},{"description":"Items per page (max 100)","in":"query","name":"limit","schema":{"default":15,"type":"integer"}},{"description":"Sort column","in":"query","name":"sort_by","schema":{"default":"created_at","enum":["created_at","updated_at","serial_number","part_condition_id"],"type":"string"}},{"description":"Sort direction","in":"query","name":"sort_dir","schema":{"default":"desc","enum":["asc","desc"],"type":"string"}},{"description":"Search keyword (matches serial_number)","in":"query","name":"q","schema":{"type":"string"}}],"responses":{"200":{"content":{"application/json":{"schema":{"allOf":[{"$ref":"#/components/schemas/SuccessResponse"},{"properties":{"data":{"properties":{"items":{"items":{"$ref":"#/components/schemas/Part"},"type":"array"},"pagination":{"$ref":"#/components/schemas/PaginationMeta"}},"type":"object"}}}]}}}}},"summary":"List parts with filtering, pagination, sorting, and search","tags":["Inventory"]},"post":{"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Part"}}},"required":true},"responses":{"201":{"content":{"application/json":{"schema":{"allOf":[{"$ref":"#/components/schemas/SuccessResponse"},{"properties":{"data":{"$ref":"#/components/schemas/Part"},"statusCode":{"example":201}}}]}}},"description":"Part created"}},"summary":"Create a part","tags":["Inventory"]}},"/inventory/parts/{id}":{"get":{"parameters":[{"in":"path","name":"id","required":true,"schema":{"format":"uuid","type":"string"}}],"responses":{"200":{"content":{"application/json":{"schema":{"allOf":[{"$ref":"#/components/schemas/SuccessResponse"},{"properties":{"data":{"$ref":"#/components/schemas/Part"}}}]}}}}},"summary":"Get part by ID","tags":["Inventory"]},"put":{"parameters":[{"in":"path","name":"id","required":true,"schema":{"format":"uuid","type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"properties":{"installation_date":{"format":"date-time","nullable":true,"type":"string"},"manufactured_date":{"format":"date-time","type":"string"},"part_catalog_id":{"format":"uuid","type":"string"},"part_condition_id":{"type":"integer"},"product_id":{"format":"uuid","nullable":true,"type":"string"},"removal_date":{"format":"date-time","nullable":true,"type":"string"},"scrapped_date":{"format":"date-time","nullable":true,"type":"string"},"serial_number":{"type":"string"}},"type":"object"}}},"required":true},"responses":{"200":{"content":{"application/json":{"schema":{"allOf":[{"$ref":"#/components/schemas/SuccessResponse"},{"properties":{"data":{"$ref":"#/components/schemas/Part"}}}]}}},"description":"Part updated"},"404":{"description":"Part not found"}},"summary":"Update part fields (partial update)","tags":["Inventory"]}},"/inventory/parts/{id}/condition":{"put":{"parameters":[{"in":"path","name":"id","required":true,"schema":{"format":"uuid","type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"properties":{"condition_id":{"type":"integer"}},"required":["condition_id"],"type":"object"}}},"required":true},"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SuccessResponse"}}},"description":"Condition updated"}},"summary":"Update part condition","tags":["Inventory"]}},"/inventory/parts/{id}/install":{"post":{"parameters":[{"in":"path","name":"id","required":true,"schema":{"format":"uuid","type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"properties":{"product_id":{"format":"uuid","type":"string"}},"required":["product_id"],"type":"object"}}},"required":true},"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SuccessResponse"}}},"description":"Part installed"}},"summary":"Install part into a product","tags":["Inventory"]}},"/inventory/parts/{id}/remove":{"post":{"parameters":[{"in":"path","name":"id","required":true,"schema":{"format":"uuid","type":"string"}}],"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SuccessResponse"}}},"description":"Part removed"}},"summary":"Remove part from product","tags":["Inventory"]}},"/inventory/parts/{id}/scrap":{"post":{"parameters":[{"in":"path","name":"id","required":true,"schema":{"format":"uuid","type":"string"}}],"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SuccessResponse"}}},"description":"Part scrapped"}},"summary":"Mark part as scrapped","tags":["Inventory"]}},"/inventory/product-models":{"post":{"requestBody":{"content":{"application/json":{"schema":{"properties":{"description":{"type":"string"},"model_code":{"type":"string"},"model_name":{"type":"string"}},"type":"object"}}},"required":true},"responses":{"201":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SuccessResponse"}}},"description":"Product model created"}},"summary":"Create a product model","tags":["Inventory"]}},"/inventory/product-models/{code}":{"get":{"parameters":[{"in":"path","name":"code","required":true,"schema":{"type":"string"}}],"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SuccessResponse"}}},"description":"Product model"}},"summary":"Get product model by code","tags":["Inventory"]}},"/inventory/products":{"get":{"parameters":[{"description":"Page number","in":"query","name":"page","schema":{"default":1,"type":"integer"}},{"description":"Items per page (max 100)","in":"query","name":"limit","schema":{"default":15,"type":"integer"}},{"description":"Sort column","in":"query","name":"sort_by","schema":{"default":"created_at","enum":["created_at","updated_at","product_name","serial_number"],"type":"string"}},{"description":"Sort direction","in":"query","name":"sort_dir","schema":{"default":"desc","enum":["asc","desc"],"type":"string"}},{"description":"Search keyword (matches product_name, serial_number)","in":"query","name":"q","schema":{"type":"string"}}],"responses":{"200":{"content":{"application/json":{"schema":{"allOf":[{"$ref":"#/components/schemas/SuccessResponse"},{"properties":{"data":{"properties":{"items":{"items":{"$ref":"#/components/schemas/Product"},"type":"array"},"pagination":{"$ref":"#/components/schemas/PaginationMeta"}},"type":"object"}}}]}}}}},"summary":"List products with pagination, sorting, and search","tags":["Inventory"]},"post":{"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Product"}}},"required":true},"responses":{"201":{"content":{"application/json":{"schema":{"allOf":[{"$ref":"#/components/schemas/SuccessResponse"},{"properties":{"data":{"$ref":"#/components/schemas/Product"},"statusCode":{"example":201}}}]}}},"description":"Product created"}},"summary":"Create a product","tags":["Inventory"]}},"/inventory/products/{id}":{"get":{"parameters":[{"in":"path","name":"id","required":true,"schema":{"format":"uuid","type":"string"}}],"responses":{"200":{"content":{"application/json":{"schema":{"allOf":[{"$ref":"#/components/schemas/SuccessResponse"},{"properties":{"data":{"$ref":"#/components/schemas/Product"}}}]}}}}},"summary":"Get product by ID","tags":["Inventory"]},"put":{"parameters":[{"in":"path","name":"id","required":true,"schema":{"format":"uuid","type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"properties":{"customer_id":{"format":"uuid","type":"string"},"product_model_code":{"type":"string"},"product_name":{"type":"string"},"serial_number":{"type":"string"}},"type":"object"}}},"required":true},"responses":{"200":{"content":{"application/json":{"schema":{"allOf":[{"$ref":"#/components/schemas/SuccessResponse"},{"properties":{"data":{"$ref":"#/components/schemas/Product"}}}]}}},"description":"Product updated"},"404":{"description":"Product not found"}},"summary":"Update product fields (partial update)","tags":["Inventory"]}},"/inventory/users/{id}":{"get":{"parameters":[{"in":"path","name":"id","required":true,"schema":{"format":"uuid","type":"string"}}],"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SuccessResponse"}}},"description":"User"}},"summary":"Get user by ID","tags":["Inventory"]}},"/purchase-orders/draft":{"post":{"requestBody":{"content":{"application/json":{"schema":{"properties":{"target_build":{"type":"string"},"vendor_id":{"format":"uuid","type":"string"}},"required":["vendor_id"],"type":"object"}}},"required":true},"responses":{"201":{"content":{"application/json":{"schema":{"allOf":[{"$ref":"#/components/schemas/SuccessResponse"},{"properties":{"data":{"$ref":"#/components/schemas/PurchaseOrder"}}}]}}},"description":"PO draft created"}},"summary":"Create a new PO draft","tags":["Purchase Orders"]}},"/purchase-orders/{poId}/approve":{"post":{"parameters":[{"in":"path","name":"poId","required":true,"schema":{"type":"string"}}],"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SuccessResponse"}}},"description":"PO approved"}},"summary":"Approve PO (transition Draft→Approved)","tags":["Purchase Orders"]}},"/purchase-orders/{poId}/line-items":{"post":{"parameters":[{"in":"path","name":"poId","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"properties":{"qty":{"minimum":1,"type":"integer"},"sku":{"type":"string"}},"required":["sku","qty"],"type":"object"}}},"required":true},"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SuccessResponse"}}},"description":"Line item added"}},"summary":"Add a line item with Eager Slot-Locking (30-min TTL)","tags":["Purchase Orders"]}},"/purchase-orders/{poId}/state":{"put":{"parameters":[{"in":"path","name":"poId","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"properties":{"new_state":{"enum":["Draft","Approved","In Transit","Received","Partial","Void"],"type":"string"}},"required":["new_state"],"type":"object"}}},"required":true},"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SuccessResponse"}}},"description":"State transitioned"}},"summary":"Transition PO state (forward only)","tags":["Purchase Orders"]}},"/shipments/{shipmentId}/dispatch":{"post":{"parameters":[{"in":"path","name":"shipmentId","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"properties":{"carrier":{"type":"string"},"operator_id":{"type":"string"},"tracking_no":{"type":"string"}},"required":["operator_id"],"type":"object"}}},"required":true},"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SuccessResponse"}}},"description":"Shipment dispatched"}},"summary":"Dispatch shipment (atomic inventory deduction)","tags":["Shipments"]}},"/shipments/{shipmentId}/lock":{"post":{"parameters":[{"in":"path","name":"shipmentId","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"properties":{"operator_id":{"type":"string"}},"required":["operator_id"],"type":"object"}}},"required":true},"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SuccessResponse"}}},"description":"Dispatch lock acquired"},"409":{"description":"Already dispatched"}},"summary":"Acquire 30-min dispatch lock","tags":["Shipments"]}},"/vendors/optimal":{"get":{"parameters":[{"in":"query","name":"sku","required":true,"schema":{"type":"string"}}],"responses":{"200":{"content":{"application/json":{"schema":{"allOf":[{"$ref":"#/components/schemas/SuccessResponse"},{"properties":{"data":{"properties":{"mapping":{"$ref":"#/components/schemas/SkuMapping"},"supplier":{"$ref":"#/components/schemas/Supplier"}},"type":"object"}}}]}}},"description":"Optimal supplier and mapping"},"404":{"description":"No optimal supplier found"}},"summary":"Route component shortage to optimal supplier","tags":["Vendors"]}},"/vendors/{id}/recalc-metrics":{"post":{"parameters":[{"in":"path","name":"id","required":true,"schema":{"format":"uuid","type":"string"}}],"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SuccessResponse"}}},"description":"Metrics updated"}},"summary":"Recalculate supplier OnTimeRate and QualityScore","tags":["Vendors"]}}},"security":[{"BearerAuth":[]}],"servers":[{"url":"https://zeus.ryanandexen.qzz.io/api/v1/scm"}]}